# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Financial Analysis Pipeline — REST API",
    "version": "1.0.0",
    "description": "REST surface of the Financial Analysis Pipeline analysis server (Express + Socket.IO on :3001). Covers connection config, LLM provider config, per-analyst source credentials / weights / flavors, market data (quote, history, options history, company news + sentiment), the Stock Screener, the multi-agent registry (agencies + analysts CRUD), and report persistence. Realtime analysis streaming itself is delivered over Socket.IO (out of scope for this REST doc).",
    "contact": { "name": "Financial Analysis Pipeline" }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local dev server" }
  ],
...
  "paths": {
...
    "/reports": {
      "post": {
        "tags": ["reports"],
        "summary": "Build + persist a report from an analysis result",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "result": { "type": "object" }, "meta": { "type": "object" } } } } } },
        "responses": {
          "200": { "description": "Persisted", "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "Missing result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Build failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "tags": ["reports"],
        "summary": "List saved reports (newest first, grouped by day)",
        "responses": {
          "200": { "description": "List", "content": { "application/json": { "schema": { "type": "object" } } } }
        }
      }
    },
    "/reports/day/{day}": {
      "get": {
        "tags": ["reports"],
        "summary": "Reports for one YYYY-MM-DD",
        "parameters": [ { "name": "day", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Day list", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/reports/{id}": {
      "get": {
        "tags": ["reports"],
        "summary": "Download a report",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["pdf", "md", "html", "json"], "default": "pdf" } }
        ],
        "responses": { "200": { "description": "File" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "tags": ["reports"],
        "summary": "Delete a report + all its generated files",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "Invalid id", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": { "type": "object", "properties": { "error": { "type": "string" }, "details": { "type": "array", "items": { "type": "string" } } } },
...
    }
  }
}